% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_distances.R
\name{view_distances}
\alias{view_distances}
\title{Looks into feature contribution (z-scores)}
\usage{
view_distances(stylo_res, pattern = "^.*?(?=_)", group = TRUE, author = NA)
}
\arguments{
\item{stylo_res}{Input data. Expects results from a `stylo()` run. Function will yell if you pass something other than 'stylo.results' class.}

\item{pattern}{String. Regex to extract class from the filename}

\item{group}{Logical. If TRUE, plots density distribution for intra- and iter-class distances separately.}

\item{author}{String. A class (e.g. author) to visualize intra-class distances.}
}
\value{
Visualizes distance distribution.
}
\description{
Compare a text or an author against corpus.
}
\examples{
\dontrun{
library(stylo)
library(seetrees)

stylo_res <- stylo(gui=F) ## run if you can
view_scores(stylo_res, target_text="CBronte_Jane")

## (reproducible) example with a `stylo` embedded dataset

data(lee) ## load the data
stylo_res <- stylo(gui=F,frequencies=lee)
view_scores(stylo_res, target_text="Faulkner_Absalom_1936")
}
}
