% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_scores.R
\name{view_scores}
\alias{view_scores}
\title{Looks into feature contribution (z-scores)}
\usage{
view_scores(stylo_res, target_text = NULL, target_class = NULL, top = NULL)
}
\arguments{
\item{stylo_res}{Input data. Expects results from a `stylo()` run. Function will yell if you pass something other than 'stylo.results' class.}

\item{target_text}{String. Filename of a document to select.}

\item{target_class}{String. A class (e.g. author) to select. In `stylo()` a class is defined by a string before the first underscore in a filename.}

\item{top}{Integer. Top distinctive features to consider}
}
\value{
Visualizes top z-scores for used features in a selected.
}
\description{
Compare a text or an author against corpus.
}
\examples{
\dontrun{
library(stylo)
library(seetrees)

stylo_res <- stylo(gui=F) ## run if you can
view_scores(stylo_res, target_text="CBronte_Jane")

## (reproducible) example with a `stylo` embedded dataset

data(lee) ## load the data
stylo_res <- stylo(gui=F,frequencies=lee)
view_scores(stylo_res, target_text="Faulkner_Absalom_1936")
}
}
