% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_tree.R
\name{view_tree}
\alias{view_tree}
\title{Cut the trees, see the words!}
\usage{
view_tree(
  stylo_res,
  k = 2,
  p = 0.05,
  color_leaves = F,
  output = T,
  label_size = 6,
  leaf_size = 1,
  right_margin = 2
)
}
\arguments{
\item{stylo_res}{Input data. Results from a stylo() run. Function will yell if you pass something other than 'stylo.results' class.}

\item{k}{Cuts a tree to a specified number of groups. Unexpected behavior possible if *k* > 5.}

\item{p}{p-value threshold for feature-to-cluster correlations. All features with p-value > p are discarded. Feel free to increase/decrease 0.05 cut-off.}

\item{color_leaves}{TRUE/FALSE. Enable coloring leaves by class (following `stylo` logic, a class is defined by a string before the first underscore in a filename)}

\item{output}{TRUE/FALSE. Enable .txt output with cluster information.}

\item{label_size}{Numeric. Controls the size of features on 'list' plots.}

\item{leaf_size}{Numeric. Controls the size of leaf labels in dendrogram.}

\item{right_margin}{Numeric. Controls the right margin of the dendrogram (use it if leaf labels do not fit). Default is 2}
}
\value{
Draws two plots and saves an optional .txt file
}
\description{
A bundled function that takes an input from a stylo() run, re-draws a dendrogram (Ward's linkage, stylo's default), cuts it to a selected numbers of groups and finds features that are associated with them.
}
\examples{
\dontrun{
library(stylo)
library(seetrees)

stylo_res <- stylo(gui=F) ## run if you can
view_tree(stylo_res, k=2) ## rebuilds a tree, cuts to k=2, calculates correlations

## (reproducible) example with stylo dataset frequencies

data(lee) ## load stylo dataset
stylo_res <- stylo(gui=F,frequencies=lee)
view_tree(stylo_res)
}
}
